Split client classes to own files
